requirements.txt
Flask==3.0.0
boto3==1.34.0
pymysql==1.1.0
python-dotenv==1.0.0
werkzeug==3.0.0
